refactor: remove basic suffix from C backend files
