+++ app/src-c/include/metalsharp_backend/diagnostics.h
