+++ app/src-c/include/metalsharp_backend/gog.h
