+++ app/src-c/runtime/antidebug.c
@@ -1,4 +1,4 @@
-#include "metalsharp_backend/antidebug_basic.h"
+#include "metalsharp_backend/antidebug.h"
 #include "metalsharp_backend/json.h"
 #include "metalsharp_backend/json_writer.h"
 #include <stdbool.h>
