+++ app/src-c/runtime/backend.c
@@ -1,36 +1,36 @@
 #include "metalsharp_backend/backend.h"
 
-#include "metalsharp_backend/antidebug_basic.h"
+#include "metalsharp_backend/antidebug.h"
 #include "metalsharp_backend/bottle_actions.h"
-#include "metalsharp_backend/bottles_basic.h"
+#include "metalsharp_backend/bottles.h"
 #include "metalsharp_backend/cache.h"
 #include "metalsharp_backend/config.h"
-#include "metalsharp_backend/d3dmetal_basic.h"
-#include "metalsharp_backend/diagnostics_basic.h"
-#include "metalsharp_backend/es_bridge_basic.h"
-#include "metalsharp_backend/game_basic.h"
-#include "metalsharp_backend/gog_basic.h"
-#include "metalsharp_backend/integration_basic.h"
+#include "metalsharp_backend/d3dmetal.h"
+#include "metalsharp_backend/diagnostics.h"
+#include "metalsharp_backend/es_bridge.h"
+#include "metalsharp_backend/game.h"
+#include "metalsharp_backend/gog.h"
+#include "metalsharp_backend/integration.h"
 #include "metalsharp_backend/json.h"
 #include "metalsharp_backend/kernel_apc.h"
 #include "metalsharp_backend/kernel_driver.h"
-#include "metalsharp_backend/kernel_extra_basic.h"
+#include "metalsharp_backend/kernel_extra.h"
 #include "metalsharp_backend/kernel_handles.h"
 #include "metalsharp_backend/kernel_integrity.h"
 #include "metalsharp_backend/logs.h"
 #include "metalsharp_backend/metalfx.h"
-#include "metalsharp_backend/migration_basic.h"
-#include "metalsharp_backend/mono_basic.h"
-#include "metalsharp_backend/mtsp_basic.h"
+#include "metalsharp_backend/migration.h"
+#include "metalsharp_backend/mono.h"
+#include "metalsharp_backend/mtsp.h"
 #include "metalsharp_backend/ob_callbacks.h"
-#include "metalsharp_backend/process_basic.h"
+#include "metalsharp_backend/process.h"
 #include "metalsharp_backend/scan.h"
 #include "metalsharp_backend/setup.h"
-#include "metalsharp_backend/sharp_basic.h"
+#include "metalsharp_backend/sharp.h"
 #include "metalsharp_backend/steam_actions.h"
-#include "metalsharp_backend/steam_basic.h"
-#include "metalsharp_backend/thread_basic.h"
-#include "metalsharp_backend/updater_basic.h"
+#include "metalsharp_backend/steam.h"
+#include "metalsharp_backend/thread.h"
+#include "metalsharp_backend/updater.h"
 
 #include <ctype.h>
 #include <errno.h>
