+++ app/src-c/runtime/d3dmetal.c
@@ -1,4 +1,4 @@
-#include "metalsharp_backend/d3dmetal_basic.h"
+#include "metalsharp_backend/d3dmetal.h"
 #include "metalsharp_backend/json.h"
 #include "metalsharp_backend/json_writer.h"
 #include <errno.h>
