+++ app/src-c/runtime/es_bridge.c
@@ -1,4 +1,4 @@
-#include "metalsharp_backend/es_bridge_basic.h"
+#include "metalsharp_backend/es_bridge.h"
 #include "metalsharp_backend/json.h"
 #include "metalsharp_backend/json_writer.h"
 #include <stdbool.h>
