+++ app/src-c/runtime/game.c
@@ -1,4 +1,4 @@
-#include "metalsharp_backend/game_basic.h"
+#include "metalsharp_backend/game.h"
 #include "metalsharp_backend/json.h"
 #include "metalsharp_backend/json_writer.h"
 #include <dirent.h>
