+++ app/src-c/runtime/integration.c
@@ -1,4 +1,4 @@
-#include "metalsharp_backend/integration_basic.h"
+#include "metalsharp_backend/integration.h"
 #include "metalsharp_backend/json.h"
 #include "metalsharp_backend/json_writer.h"
 #include <stdbool.h>
