+++ app/src-c/runtime/migration.c
@@ -1,4 +1,4 @@
-#include "metalsharp_backend/migration_basic.h"
+#include "metalsharp_backend/migration.h"
 #include "metalsharp_backend/json.h"
 #include "metalsharp_backend/json_writer.h"
 #include "metalsharp_backend/setup.h"
