+++ app/src-c/runtime/mtsp.c
@@ -1,4 +1,4 @@
-#include "metalsharp_backend/mtsp_basic.h"
+#include "metalsharp_backend/mtsp.h"
 
 #include "metalsharp_backend/json.h"
 #include "metalsharp_backend/json_writer.h"
