+++ app/src-c/runtime/mtsp_catalog.c
@@ -1,4 +1,4 @@
-#include "metalsharp_backend/mtsp_basic.h"
+#include "metalsharp_backend/mtsp.h"
 #include <stdlib.h>
 #include <string.h>
