+++ app/src-c/runtime/process.c
@@ -1,4 +1,4 @@
-#include "metalsharp_backend/process_basic.h"
+#include "metalsharp_backend/process.h"
 #include "metalsharp_backend/json.h"
 #include "metalsharp_backend/json_writer.h"
 #include <dirent.h>
