+++ app/src-c/runtime/steam.c
@@ -1,4 +1,4 @@
-#include "metalsharp_backend/steam_basic.h"
+#include "metalsharp_backend/steam.h"
 
 #include "metalsharp_backend/json.h"
 #include "metalsharp_backend/json_writer.h"
