+++ app/src-c/runtime/thread.c
@@ -1,4 +1,4 @@
-#include "metalsharp_backend/thread_basic.h"
+#include "metalsharp_backend/thread.h"
 #include "metalsharp_backend/json.h"
 #include "metalsharp_backend/json_writer.h"
 #include <stdbool.h>
