+++ app/src-c/runtime/updater.c
@@ -1,4 +1,4 @@
-#include "metalsharp_backend/updater_basic.h"
+#include "metalsharp_backend/updater.h"
 
 #include "metalsharp_backend/backend.h"
 #include "metalsharp_backend/json.h"
